Moved QueryGenerators to their own folder
