+++ aspnetcore/src/Api/Services/ElasticSearchQueryGenerators/IQueryGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using Nest;
 
-namespace Api.Services
+namespace Api.Services.ElasticSearchQueryGenerators
 {
     public interface IQueryGenerator<TIn, TOut> where TOut : class
     {
